Create diceRoll.cs
Using the Random.Next() method to simulate rolling three six-sided dice. We'll evaluate the values to calculate the score. If the score is greater than an arbitrary total, we'll display a winning message to the user. Otherwise, we'll display a losing message to the user.

If any two dice you roll result in the same value, you get two bonus points for rolling doubles.
If all three dice you roll result in the same value, you get six bonus points for rolling triples.
If the sum of the three dice rolls, plus any point bonuses, is 15 or greater, you win the game. Otherwise, you lose.