Create discountExercise
Code challenges throughout these modules reinforce what you've learned and help you gain some confidence before proceeding.

Challenge: Improve renewal rate of subscriptions
You've been asked to add a feature to your company's software. The feature is intended to improve the renewal rate of subscriptions to the software. Your task is to display a renewal message when a user signs in to the software system and is notified that their subscription will soon end. You'll need to add a couple of decision statements to properly add branching logic to the application to satisfy the requirements.

Rule 1. If the user's subscription will expire in 10 days or less, display the message:

Output
Your subscription will expire soon. Renew now!
Rule 2. If the user's subscription will expire in 5 days or less, display the messages:

Output
Your subscription expires in x days.
Renew now and save 10%!
Make sure to substitute x for the value stored in the variable daysUntilExpiration.

Rule 3. If the user's subscription will expire in 1 day, display the message:

Output


Your subscription expires within a day!
Renew now and save 20%!